Update "AutoCompleteManager" to be robust against null values
When using the "autocomplete" field in one of our templates, we learned, that there are some cases where the internal variables of the "AutoCompleteManager" - such as "_textBox", "_scrollViewer", "_resizeGrip", "_scrollBar" - can be null in some edge cases. I've yet to analyze why that happens sometimes, but i figured it would be a good idea to add some null-checks when using those variables, before this component throws some "NullReferenceException"s that the consumer has to handle.